heater_pc: define predictive control extras
Allows for the amendment of the real control loop
with feed-forward control

Feed-forward allows for the implementation of advanced control,
such as "dual loop", using data from several temperature sensors.
Increase/decrease power beforehand without PID reactive lag.

Templating happens in the reactor,
to make access to the status fields thread-safe.

Template tested on "connect",
So it will crash early and not inside the timer.

Signed-off-by: Timofey Titovets <nefelim4ag@gmail.com>